Fix duplicate keys issue + use camelCase for all keys
There were multiple cases where a key could be renamed but the original
would not be deleted, which resulted in stats section double the size
they should be. This is now fixed.

During this process, we switched to camelCase-ing those names instead of
using a custom string substitution. Which means it is a bit less close
to what is displayed on playoverwatch's profile page, but is easier to
read (at least for a dev, and potential end users probably would not
care for either the earlier substitution, nor the latter, but rather a
nice GUI in front of it)